move icds_pre_release_features
