+++ custom/icds_reports/utils/__init__.py
@@ -28,12 +28,11 @@ from openpyxl import Workbook
 from openpyxl.utils import get_column_letter
 from weasyprint import HTML, CSS
 
-from corehq import toggles
 from corehq.apps.app_manager.dbaccessors import get_latest_released_build_id
 from corehq.apps.locations.models import SQLLocation
 from corehq.apps.reports.datatables import DataTablesColumn
 from corehq.apps.reports_core.filters import Choice
-from corehq.apps.userreports.models import StaticReportConfiguration, AsyncIndicator
+from corehq.apps.userreports.models import StaticReportConfiguration
 from corehq.apps.userreports.reports.data_source import ConfigurableReportDataSource
 from corehq.blobs.mixin import safe_id
 from corehq.const import ONE_DAY
@@ -51,15 +50,13 @@ from custom.icds_reports.const import (
     THR_21_DAYS_THRESHOLD_DATE
 )
 
-from custom.icds_reports.exceptions import InvalidLocationTypeException
 from custom.icds_reports.models.aggregate import AwcLocation
 from custom.icds_reports.models.helper import IcdsFile
 from custom.icds_reports.queries import get_test_state_locations_id, get_test_district_locations_id
 from couchexport.export import export_from_tables
 from dimagi.utils.dates import DateSpan
 from dimagi.utils.parsing import ISO_DATE_FORMAT
-from django.db.models import Case, When, Q, F, IntegerField, Max, Min
-from django.db.utils import OperationalError
+from django.db.models import Case, When, Q, F, IntegerField, Min
 import uuid
 from sqlagg.filters import EQ, NOT
 from pillowtop.models import KafkaCheckpoint
@@ -760,10 +757,6 @@ def generate_qrcode(data):
     image.save(output, "PNG")
     qr_content = b64encode(output.getvalue())
     return qr_content
-
-
-def icds_pre_release_features(user):
-    return toggles.ICDS_DASHBOARD_REPORT_FEATURES.enabled(user.username)
 
 
 def indian_formatted_number(number):
